pyproject.toml
[project]
name = "class-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents[litellm]>=0.0.13",
    "python-dotenv>=1.1.0",
]
